flask_app/requirements.txt
flask
uWSGI
pymongo
PyYAML
flask_pymongo